+++ ticketing/views/__init__.py
@@ -1,10 +1,7 @@
-from .remains import (
-    delete_ticket_item,
-    index,
-    set_ticket_to_event,
-    ticket_item_edit,
-)
+from .buy_tickets import index
+from .set_ticket_to_event import set_ticket_to_event
 from .ticket_event_edit import ticket_event_edit
 from .ticket_items import ticket_items
 from .transactions import transactions
 from .create_transaction import CreateTransaction
+from .ticket_item_edit import ticket_item_edit
